+++ src/app/auth/error.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import type { NextPage } from 'next';
+
+const Error: NextPage = (props) => {
+  return <>Login Error</>;
+};
+
+export default Error;
