+++ src/components/BackButton.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React from 'react';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 
 const BackButton = () => {
   const router = useRouter();
