+++ src/libs/hooks/services/mutations/member.ts
@@ -1,6 +1,6 @@
-import {useFetchWrapper} from '@libs/client/fetch-wrapper';
-import {useMutation} from '@tanstack/react-query';
-import {useRouter} from 'next/router';
+import { useFetchWrapper } from '@libs/client/fetch-wrapper';
+import { useMutation } from '@tanstack/react-query';
+import { useRouter } from 'next/navigation';
 
 interface UpdateForm {
   email: string;
@@ -16,13 +16,13 @@ const memberKeys = {
 
 //회원정보 수정(public profile)
 function usePublicProfileUpdate(): any {
-  const {patch} = useFetchWrapper();
+  const { patch } = useFetchWrapper();
   const router = useRouter();
 
   return useMutation({
     mutationKey: memberKeys.update('publicProfile'),
     mutationFn: (value: UpdateForm) =>
-      patch({url: `/member/setting/info/${value.email}`, data: value}),
+      patch({ url: `/member/setting/info/${value.email}`, data: value }),
     onMutate: (variables) => {},
     onSuccess: (data, variables, context) => {
       return router.push('/profile');
@@ -34,12 +34,12 @@ function usePublicProfileUpdate(): any {
 
 //회원정보 수정(account)
 function useAccountSave(): any {
-  const {post} = useFetchWrapper();
+  const { post } = useFetchWrapper();
   const router = useRouter();
 
   return useMutation({
     mutationKey: memberKeys.update('saveAccount'),
-    mutationFn: (value: UpdateForm) => post({url: `/member/account`, data: value}),
+    mutationFn: (value: UpdateForm) => post({ url: `/member/account`, data: value }),
     onMutate: (variables) => {},
     onSuccess: (data, variables, context) => {
       return router.push('/profile');
@@ -50,12 +50,13 @@ function useAccountSave(): any {
 }
 
 function useAccountInfoUpdate(): any {
-  const {patch} = useFetchWrapper();
+  const { patch } = useFetchWrapper();
   const router = useRouter();
 
   return useMutation({
     mutationKey: memberKeys.update('temp'),
-    mutationFn: (value: UpdateForm) => patch({url: `/member/account/${value.email}`, data: value}),
+    mutationFn: (value: UpdateForm) =>
+      patch({ url: `/member/account/${value.email}`, data: value }),
     onMutate: (variables) => {},
     onSuccess: (data, variables, context) => {
       // return router.push('/profile');
@@ -65,4 +66,4 @@ function useAccountInfoUpdate(): any {
   });
 }
 
-export {usePublicProfileUpdate, useAccountInfoUpdate, useAccountSave};
+export { usePublicProfileUpdate, useAccountInfoUpdate, useAccountSave };
