+++ src/libs/hooks/services/mutations/order.ts
@@ -1,18 +1,18 @@
-import {useFetchWrapper} from '@libs/client/fetch-wrapper';
-import {useMutation, useQueryClient} from '@tanstack/react-query';
-import {orderKeys} from '@libs/hooks/services/keys/order';
-import {IItem, IRecruit, IRecruitJoin} from '../../../../types/order';
-import {router} from 'next/client';
-import {useRouter} from 'next/router';
-import {useSession} from 'next-auth/react';
+import { useFetchWrapper } from '@libs/client/fetch-wrapper';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { orderKeys } from '@libs/hooks/services/keys/order';
+import { IItem, IRecruit, IRecruitJoin } from '../../../../types/order';
+import { router } from 'next/client';
+import { useRouter } from 'next/navigation';
+import { useSession } from 'next-auth/react';
 
 function useOrderStartRecruit(): any {
-  const {post} = useFetchWrapper();
+  const { post } = useFetchWrapper();
   const cache = useQueryClient();
 
   return useMutation({
     mutationKey: orderKeys.insert(),
-    mutationFn: (value: IRecruit) => post({url: `/order/recruit`, data: value}),
+    mutationFn: (value: IRecruit) => post({ url: `/order/recruit`, data: value }),
     onMutate: (variables) => {},
     onSuccess: (data, variables, context) => {},
     onSettled: (data, error, variables, context) => {
@@ -23,13 +23,13 @@ function useOrderStartRecruit(): any {
 }
 
 function useOrderRecruitGroupJoin(): any {
-  const {post} = useFetchWrapper();
+  const { post } = useFetchWrapper();
   const cache = useQueryClient();
   const router = useRouter();
 
   return useMutation({
     mutationKey: orderKeys.insert(),
-    mutationFn: (value: IRecruitJoin) => post({url: `/order/recruit/group/join`, data: value}),
+    mutationFn: (value: IRecruitJoin) => post({ url: `/order/recruit/group/join`, data: value }),
     onMutate: (variables) => {},
     onSuccess: (data, variables, context) => {
       router.reload();
@@ -42,14 +42,14 @@ function useOrderRecruitGroupJoin(): any {
 }
 
 function useOrderRecruitGroupExit(): any {
-  const {axiosDelete} = useFetchWrapper();
+  const { axiosDelete } = useFetchWrapper();
   const cache = useQueryClient();
   const router = useRouter();
 
   return useMutation({
     mutationKey: orderKeys.insert(),
     mutationFn: (groupOrderId: number) =>
-      axiosDelete({url: `/order/recruit/group/join/${groupOrderId}`}),
+      axiosDelete({ url: `/order/recruit/group/join/${groupOrderId}` }),
     onMutate: (variables) => {},
     onSuccess: (data, variables, context) => {
       router.reload();
