--- src/pages/auth/error.tsx
@@ -1,7 +1,0 @@
-import type {NextPage} from 'next';
-
-const Error: NextPage = (props) => {
-  return <>Login Error</>;
-};
-
-export default Error;
